Move vision backend API to its own app.
In the future, if we commit more to an API-driven project design, we may want to make vision_backend hold the models, utils, and API views, while a new app called vision_backend_web holds the HTML-served views. This way the models don't have to migrate their app name. However, we're just getting started with the API now, so this separate VB-api app seems more appropriate for the time being.