Add a hidden flag to Jobs, and apply this to certain cases of check_source, run_scheduled_jobs, collect_spacer_jobs, and schedule_periodic_jobs.
Indicate this with ^ on job lists, and move that and * from Timing-info to Type. Also add a field to actually hide these jobs or not.

The code to decide what persists and what's hidden is located close to each task, instead of all being in finish_job().

A little bit of test utils refactoring.