collect_spacer_jobs(): Collect jobs in pk order (FIFO).
Normally this doesn't make a ton of difference, except when the server is overloaded enough that collect_spacer_jobs() consistently times out - as a result of both high spacer job volume, and this job only being able to run a couple times a day. In that case, some spacer jobs might remain uncollected indefinitely.