Use Events, not django-reversion, to track machine-classification actions for annotation history from this point on.
django-reversion objects will still be used to display previous machine-classification actions (as well as all other types of annotation actions) for the time being. But from here on out, django-reversion no longer creates new records for machine classification.

While we're at it, since bulk_create() is now safe to use for creating annotations during machine classification, that's exactly what we do to get a speedup. At least in the common case where we're creating annotations for all points of an image.