Upload images: after upload, have the metadata-edit link use an image ID range parameter instead of a list of IDs parameter.
This ensures that the parameter length doesn't grow unbounded as more images are uploaded at once.

Also have this and "Upload more images" be GET links instead of POST submissions.

Note: Image ID range is not foolproof, since it can get images outside of the upload set if there are concurrent uploads in the same source. But not sure if the problem is big enough to warrant crafting a more complex solution.