More parallel structure between Django tests of several browse-filtering views: images, metadata, patches, annotation tool, delete images, and delete annotations.
No more filtering tests for most non-aux metadata fields, because those fields were never rendered on the search form in the first place. (They were in the Django form's fields, but not in the fieldsets attribute that is used for rendering. They've now been removed from the former.)

Landing on browse and then going to page 2 no longer counts as "searched or filtered" for the purposes of delete eligibility.