Batch delete images/annotations: compare the deletion count with the search result count to double check, and abort if they don't match.
We have a true "time of deletion" count for delete images, since it uses the number returned by the deletion call. But for delete annotations, the count is taken just *before* deleting annotations. I couldn't think of a way to make the latter also a "true" count without greatly sacrificing performance.