Browse search form: define a limit for the aux. metadata dropdown length, and use a text field instead when the option count would exceed that length.
This helps avoid the scenario where a metadata dropdown field gets so long that it slows down the entire page's interactivity.

Add help text explaining this behavior as well as giving related tips.

While we're at it, in the "every image has the same non-blank aux. meta value" case, render the aux. meta field anyway. Hiding the field in this case was a bit too clever and non-obvious, even if it did have some sense to it.