Browse Patches: only count up to 20000 results.
This means that, in a situation where the paginator would previously count 80000 results and give 4000 pages, we'd now only say "Showing m-n of 20000 or more" and give only up to 1000 pages.

This is implemented by having paginator.count call `self.object_list[:20000].count()` instead of `self.object_list.count()`. This apparently allows the SQL count query to stop once it reaches 20000, thus giving us a speedup when there would be many more results in total.

Note that the count can still be very slow if many database rows have to be scanned to get to 20000 matching results. So having the right database indexes is still important.